gateway/requirements.txt
fastapi==0.115.12
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
uvicorn==0.34.2
pydantic-settings==2.9.1
psycopg2-binary==2.9.10
httpx==0.28.1
# jwt==1.3.1
PyJWT==2.10.1
python-multipart==0.0.20